+++ src/main/java/com/github/siboxd/fatturapa/model/DatiTrasportoType.java
@@ -1,5 +1,7 @@
 package com.github.siboxd.fatturapa.model;
 
+import com.github.siboxd.fatturapa.model.invoiceheader.supplier.IndirizzoType;
+
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
