+++ src/main/java/com/github/siboxd/fatturapa/model/invoiceheader/supplier/CedentePrestatoreType.java
@@ -1,11 +1,14 @@
-package com.github.siboxd.fatturapa.model;
+package com.github.siboxd.fatturapa.model.invoiceheader.supplier;
+
+import com.github.siboxd.fatturapa.model.ContattiType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
 
 
 /**
- * CedentePrestatoreType<br>
+ * This block contains the data relating to the seller of the good / service being billed.
+ * <p>
  * Generated using Android JAXB<br>
  *
  * @link https://github.com/yeshodhan/android-jaxb
@@ -13,16 +16,21 @@ import org.simpleframework.xml.Root;
 @Root(name = "CedentePrestatoreType")
 public class CedentePrestatoreType {
 
-    @Element(name = "DatiAnagrafici", required = true)
+    @Element(name = "DatiAnagrafici")
     private DatiAnagraficiCedenteType datiAnagrafici;
-    @Element(name = "Sede", required = true)
+
+    @Element(name = "Sede")
     private IndirizzoType sede;
+
     @Element(name = "StabileOrganizzazione", required = false)
     private IndirizzoType stabileOrganizzazione;
+
     @Element(name = "IscrizioneREA", required = false)
     private IscrizioneREAType iscrizioneREA;
+
     @Element(name = "Contatti", required = false)
     private ContattiType contatti;
+
     @Element(name = "RiferimentoAmministrazione", required = false)
     private String riferimentoAmministrazione;
 
@@ -33,7 +41,7 @@ public class CedentePrestatoreType {
         return datiAnagrafici;
     }
 
-    public void setDatiAnagrafici(DatiAnagraficiCedenteType datiAnagrafici) {
+    public void setDatiAnagrafici(final DatiAnagraficiCedenteType datiAnagrafici) {
         this.datiAnagrafici = datiAnagrafici;
     }
 
@@ -41,7 +49,7 @@ public class CedentePrestatoreType {
         return sede;
     }
 
-    public void setSede(IndirizzoType sede) {
+    public void setSede(final IndirizzoType sede) {
         this.sede = sede;
     }
 
@@ -49,7 +57,12 @@ public class CedentePrestatoreType {
         return stabileOrganizzazione;
     }
 
-    public void setStabileOrganizzazione(IndirizzoType stabileOrganizzazione) {
+    /**
+     * <b>Note:</b> To be valued only if the supplier is a person who does not reside in Italy
+     * but who, in Italy, has a permanent establishment through which he carries out his business
+     * (he makes sales of goods or services that are billed)
+     */
+    public void setStabileOrganizzazione(final IndirizzoType stabileOrganizzazione) {
         this.stabileOrganizzazione = stabileOrganizzazione;
     }
 
@@ -57,7 +70,12 @@ public class CedentePrestatoreType {
         return iscrizioneREA;
     }
 
-    public void setIscrizioneREA(IscrizioneREAType iscrizioneREA) {
+    /**
+     * <b>Note:</b> To be valued only if the seller is a company registered in the business register
+     * and as such has the obligation to indicate in all the documents also the registration data
+     * (Article 2250 of the Italian Civil Code)
+     */
+    public void setIscrizioneREA(final IscrizioneREAType iscrizioneREA) {
         this.iscrizioneREA = iscrizioneREA;
     }
 
@@ -65,7 +83,7 @@ public class CedentePrestatoreType {
         return contatti;
     }
 
-    public void setContatti(ContattiType contatti) {
+    public void setContatti(final ContattiType contatti) {
         this.contatti = contatti;
     }
 
@@ -73,7 +91,7 @@ public class CedentePrestatoreType {
         return riferimentoAmministrazione;
     }
 
-    public void setRiferimentoAmministrazione(String riferimentoAmministrazione) {
+    public void setRiferimentoAmministrazione(final String riferimentoAmministrazione) {
         this.riferimentoAmministrazione = riferimentoAmministrazione;
     }
